Fixed build warnings
The repo no builds with no warnings. Let's keep it as such.

+++ src/System.Binary/System/Binary/BigEndianReader.cs
@@ -32,7 +32,7 @@ namespace System.Binary.BigEndian
         {
             short result = data[0];
             result <<= 8;
-            result |= data[1];
+            result |= (short)data[1];
             return result;
         }
 
@@ -40,7 +40,7 @@ namespace System.Binary.BigEndian
         {
             short result = data[0];
             result <<= 8;
-            result |= data[1];
+            result |= (short)data[1];
             return result;
         }
